refactor: PluginManagePage 迁入 com.tungsten.fcl.ui.setting 包
与 LauncherSettingPage/AboutPage 等设置页子页面同包，删除 SettingUI 的同包 import
